feat: 베팅 머니 구현 완료
