+++ src/test/java/domain/BettingMoneyTest.java
@@ -12,16 +12,16 @@ import org.junit.jupiter.params.provider.ValueSource;
 /**
  * @author 우가
  * @version 1.0.0
- * @Created by 우가 on 2023/03/10
+ * @Created by 우가 on 2023/03/11
  */
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public class MoneyTest {
+public class BettingMoneyTest {
 
     @ParameterizedTest
     @ValueSource(ints = {0, 99, -100})
     void 백원이하를_넘겨주면_예외가_발생한다(int money) {
-        assertThatThrownBy(() -> new Money(money))
+        assertThatThrownBy(() -> new BettingMoney(money))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("최소 베팅 금액은 100원 입니다.");
     }
@@ -29,7 +29,7 @@ public class MoneyTest {
     @ParameterizedTest
     @ValueSource(ints = {101, 1001})
     void 백원단위가_아닐경우_예외가_발생한다(int money) {
-        assertThatThrownBy(() -> new Money(money))
+        assertThatThrownBy(() -> new BettingMoney(money))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("100원 단위로 입력가능합니다.");
     }
@@ -37,14 +37,14 @@ public class MoneyTest {
     @ParameterizedTest
     @ValueSource(ints = {100_000_100, 1_000_000_000})
     void 베팅_금액은_일억을_넘길경우_예외가_발생한다(int money) {
-        assertThatThrownBy(() -> new Money(money))
+        assertThatThrownBy(() -> new BettingMoney(money))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("최대 베팅 금액은 1억원 입니다.");
     }
 
     @Test
     void 수익률에_따른_베팅_금액을_반환한다() {
-        Money money = new Money(1000);
+        BettingMoney money = new BettingMoney(1000);
 
         double profit = 1.5;
